useMemo not in z view
