+++ static/app/views/releases/utils/useFetchThresholdsListData.tsx
@@ -1,7 +1,11 @@
+import {useMemo} from 'react';
+
+import type {Environment} from 'sentry/types';
 import {MonitorType} from 'sentry/types/alerts';
 import {type ApiQueryKey, useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
 import type {MetricRule} from 'sentry/views/alerts/rules/metric/types';
+import {TOTAL_ERROR_COUNT_STR} from 'sentry/views/releases/utils/constants';
 
 import type {Threshold, ThresholdQuery} from './types';
 
@@ -47,8 +51,36 @@ export default function useFetchThresholdsListData({
     organization.features?.includes('releases-v2') ||
     organization.features?.includes('releases-v2-st');
 
-  return useApiQuery<Threshold[] | MetricRule[]>(queryKey, {
+  const {data, ...remainigProps} = useApiQuery<Threshold[] | MetricRule[]>(queryKey, {
     staleTime: 0,
     enabled: hasReleaseV2Feature || isActivatedAlert,
   });
+
+  const thresholds: Threshold[] = useMemo(() => {
+    if (!isActivatedAlert) return thresholds;
+    if (!data) return [];
+
+    return data.map((thresholdOrRule: Threshold | MetricRule) => {
+      const rule = thresholdOrRule as MetricRule;
+
+      return {
+        date_added: rule.dateCreated,
+        environment: {
+          name: rule.environment,
+          displayName: rule.environment,
+        } as Environment,
+        id: rule.id,
+        project: {id: rule.projects[0], slug: rule.projects[0]},
+        threshold_type: TOTAL_ERROR_COUNT_STR,
+        trigger_type: 'over',
+        value: rule.triggers[0]?.alertThreshold,
+        window_in_seconds: rule.timeWindow,
+      } as Threshold;
+    });
+  }, [isActivatedAlert, data]);
+
+  return {
+    data: thresholds,
+    ...remainigProps,
+  };
 }
